+++ src/features/defi/helpers/normalizeOpportunity.tsx
@@ -68,18 +68,18 @@ const useTransformCosmosStaking = (
 
 type NormalizeOpportunitiesProps = {
   cosmosSdkStakingOpportunities: MergedActiveStakingOpportunity[]
-  foxEthLpOpportunity?: EarnOpportunityType
   stakingOpportunities?: EarnOpportunityType[]
+  lpOpportunities?: EarnOpportunityType[]
 }
 
 export const useNormalizeOpportunities = ({
   cosmosSdkStakingOpportunities = [],
-  foxEthLpOpportunity,
   stakingOpportunities,
+  lpOpportunities,
 }: NormalizeOpportunitiesProps): EarnOpportunityType[] => {
   return [
     ...useTransformCosmosStaking(cosmosSdkStakingOpportunities),
     ...(stakingOpportunities ? stakingOpportunities : []),
-    ...(foxEthLpOpportunity ? [foxEthLpOpportunity] : []),
+    ...(lpOpportunities ? lpOpportunities : []),
   ]
 }
