+++ src/pages/Defi/hooks/useEarnBalances.tsx
@@ -6,7 +6,7 @@ import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { useCosmosSdkStakingBalances } from 'pages/Defi/hooks/useCosmosSdkStakingBalances'
 import { foxEthLpAssetId } from 'state/slices/opportunitiesSlice/constants'
 import {
-  selectAggregatedEarnUserLpOpportunity,
+  selectAggregatedEarnUserLpOpportunities,
   selectAggregatedEarnUserStakingOpportunities,
   selectPortfolioFiatBalanceByAssetId,
 } from 'state/slices/selectors'
@@ -35,12 +35,7 @@ export function useEarnBalances(): UseEarnBalancesReturn {
 
   const stakingOpportunities = useAppSelector(selectAggregatedEarnUserStakingOpportunities)
 
-  const foxEthLpOpportunity = useAppSelector(state =>
-    selectAggregatedEarnUserLpOpportunity(state, {
-      lpId: foxEthLpAssetId,
-      assetId: foxEthLpAssetId,
-    }),
-  )
+  const lpOpportunities = useAppSelector(selectAggregatedEarnUserLpOpportunities)
 
   const stakingContractsAggregatedOpportunities = useAppSelector(
     selectAggregatedEarnUserStakingOpportunities,
@@ -61,6 +56,7 @@ export function useEarnBalances(): UseEarnBalancesReturn {
     }),
     [],
   )
+
   const foxEthLpFiatBalance = useAppSelector(state =>
     selectPortfolioFiatBalanceByAssetId(state, lpAssetBalanceFilter),
   )
@@ -69,7 +65,7 @@ export function useEarnBalances(): UseEarnBalancesReturn {
     cosmosSdkStakingOpportunities: cosmosSdkStakingOpportunities.concat(
       osmosisStakingOpportunities,
     ),
-    foxEthLpOpportunity,
+    lpOpportunities,
     stakingOpportunities,
   })
